Caches for Recipes & Ingredients
Added a GenericCache to reduce hot calls to mongodb. The cache expects the objects to have a value of _id. This is so that in the dict it can use this as the key.

On CRUD of an item (Recipe/ingredient) it will then update the cache with the updated values.

The cache is only initialised after the first call to the database when the app runs. This is so that if the app crashes, there isn't a malformed cache.