ApiResponseHandling and event propagation
Added response types for the current available returned types from mongo for the responses. This then allows a check on the front end to display back to the user when the update has completed.

Added a mousedown event to the recipe step cards to stop parent events from affecting them. This is because with CDK drag the text inputs for steps weren't easily accessible. This ensures when you click in edit mode that the cursor to type appears.